requirements.txt
requests==2.24.0
six==1.15.0
Werkzeug==1.0.1
Flask==1.1.2
flask_wtf==0.14.3
protobuf==3.14.0
pyrebase==3.0.27
WTForms==2.3.3
google-cloud-firestore==1.8.1
google-cloud-storage==1.29.0
google-cloud-logging==1.15.0
google-cloud-error-reporting==0.34.0
gunicorn==20.0.4
six==1.15.0